std: move svs defns
