+++ std/ndn/svs/v3/definitions.go
@@ -29,16 +29,6 @@ type SeqNoEntry struct {
 	SeqNo uint64 `tlv:"0xd6"`
 }
 
-// This actually belongs in SVS-PS but codegen doesn't support cross-package
-type InstanceState struct {
-	//+field:name
-	Name enc.Name `tlv:"0x07"`
-	//+field:natural
-	BootstrapTime uint64 `tlv:"0xd4"`
-	//+field:struct:StateVector
-	StateVector *StateVector `tlv:"0xc9"`
-}
-
 // +tlv-model:nocopy
 type PassiveState struct {
 	//+field:sequence:[]byte:binary:[]byte
